pyproject.toml
[project]
name = "reachy-local-agent"
version = "0.1.0"
description = "Local voice-driven agent for Reachy Mini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "reachy-mini>=1.6.0",
    "faster-whisper>=1.1.0",
    "piper-tts>=1.4.0",
    "openai>=1.0.0",
    "silero-vad>=5.1",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "opencv-contrib-python>=4.10.0",
]

[project.optional-dependencies]
sim = ["reachy-mini[mujoco]"]
dances = ["reachy-mini-dances-library>=0.2.0"]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
    "ty>=0.0.1a25",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I"]
